reinforcement-learning/RL_PACKAGE/requirements.txt
category_encoders==2.6.3
gymnasium==0.29.1
networkx==3.2.1
numpy==1.26.4
pandas==2.2.1
pyintergraph==1.3.3
igraph==0.10.6
sb3_contrib==2.2.1
scikit_learn==1.4.1.post1
stable_baselines3==2.2.1
torch_geometric==2.5.0
tqdm==4.66.2
umap-learn==0.5.5
pyg_lib
torch_scatter
torch_sparse
torch==2.1.0

